pyproject.toml
[tool.poetry]
name = "nonebot_plugin_tarot"
version = "0.3.0a1"
description = "Tarot divination!"
authors = ["KafCoppelia <k740677208@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.2"
nonebot-adapter-onebot = "^2.0.0-beta.1"
pillow = "^9.0.0"
aiocache = "^0.11.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tencentyun"
url = "https://mirrors.tencent.com/pypi/simple"
default = true